Travis: do not cache the downloads dir.
The downloads dir is only created when buildout:download-cache is set in ~/.buildout/default.cfg,
or if it is passed on the command line.
Seems unneeded, because we are only interested in the resulting eggs.
If buildout downloads non-eggs, for example a varnish tarball, then caching the downloads dir could be useful.